Refactor: Split adapter into one adapter for each service
One adapter each for Diagnostics, Resource, Data instead of one
to ease composition of handlers.
