+++ backend/data.go
@@ -1,0 +1,42 @@
+package backend
+
+import (
+	"context"
+	"encoding/json"
+	"time"
+
+	"github.com/grafana/grafana-plugin-sdk-go/data"
+)
+
+type QueryDataRequest struct {
+	PluginConfig PluginConfig
+	Headers      map[string]string
+	Queries      []DataQuery
+	User         *User
+}
+
+// DataQuery represents the query as sent from the frontend.
+type DataQuery struct {
+	RefID         string
+	MaxDataPoints int64
+	Interval      time.Duration
+	TimeRange     TimeRange
+	JSON          json.RawMessage
+}
+
+// QueryDataResponse holds the results for a given query.
+type QueryDataResponse struct {
+	Frames   []*data.Frame
+	Metadata map[string]string
+}
+
+// TimeRange represents a time range for a query.
+type TimeRange struct {
+	From time.Time
+	To   time.Time
+}
+
+// QueryDataHandler handles data queries.
+type QueryDataHandler interface {
+	QueryData(ctx context.Context, req *QueryDataRequest) (*QueryDataResponse, error)
+}
